requirements.txt
fastapi
uvicorn
openai
fastapi-clerk-auth
python-dotenv